+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Simple autodetect of game libraries
 * eglGetProcAddress() is also used for gl functions
 * Prevent a mono sleep from adding variable time to the timer
+* Reorder codecs and add Ut (#467)
 
 ### Changed